ADD TargetDealDamageToPlayerBroken
It seems that I have somehow broken the code for the Target damaging the Player, now when the Target fires no damage is dealt. Originally it dealt damage to the player when a projectile was fired rather than when it hit the player. I tried to fix it and now it does no damage at all