ADD TodayFinalTweaks - 09/12/24
When I do come back to this project, I think my main plan is to add as much as I can to the Level, mainly to the exteriors of the buildings, as right now they are all cubes. I think adding more detail to the outside would make the level more visually pleasing.

Also need to finish up the interiors and the surrounding landscape of the level, then add the final targets, then just final tweaks and fixes and we should be good to go!!!